Flake8 fixups
