+++ daisy_workflows/linux_common/utils/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/env python2.7
+
+from common import *  # noqa: F401,F403
+from diskutils import *  # noqa: F401,F403
