+++ daisy_workflows/linux_common/utils/common.py
@@ -29,7 +29,6 @@ import traceback
 import urllib2
 import uuid
 
-
 SUCCESS_LEVELNO = logging.ERROR - 5
 
 
@@ -113,7 +112,7 @@ def GetCurrentLoginProfileUsername(user_lib, unique_id_user):
   Equivalent of calling the gcloud equivalent:
 
   gcloud compute os-login describe-profile --format \
-      value\(posixAccounts.username\)
+      value\\(posixAccounts.username\\)
 
   Parameter:
   Args:
@@ -136,61 +135,6 @@ def GetServiceAccountUniqueIDUser():
   s = _GetMetadataParam('service-accounts/default/?recursive=True')
   service_info = json.loads(s)
   return 'users/' + service_info['email']
-
-
-def MountDisk(disk):
-  # Note: guestfs is not imported in the beginning of the file as it might not
-  # be installed when this module is loaded
-  import guestfs
-
-  # All new Python code should pass python_return_dict=True
-  # to the constructor.  It indicates that your program wants
-  # to receive Python dicts for methods in the API that return
-  # hashtables.
-  g = guestfs.GuestFS(python_return_dict=True)
-  # Set the product name as cloud-init checks it to confirm this is a VM in GCE
-  g.config('-smbios', 'type=1,product=Google Compute Engine')
-  g.set_verbose(1)
-  g.set_trace(1)
-
-  g.set_memsize(4096)
-
-  # Enable network
-  g.set_network(True)
-
-  # Attach the disk image to libguestfs.
-  g.add_drive_opts(disk)
-
-  # Run the libguestfs back-end.
-  g.launch()
-
-  # Ask libguestfs to inspect for operating systems.
-  roots = g.inspect_os()
-  if len(roots) == 0:
-    raise Exception('inspect_vm: no operating systems found')
-
-  # Sort keys by length, shortest first, so that we end up
-  # mounting the filesystems in the correct order.
-  mps = g.inspect_get_mountpoints(roots[0])
-
-  def compare(a, b):
-    return len(a) - len(b)
-
-  for device in sorted(mps.keys(), compare):
-    try:
-      g.mount(mps[device], device)
-    except RuntimeError as msg:
-      logging.warn('%s (ignored)' % msg)
-
-  return g
-
-
-def UnmountDisk(g):
-  try:
-    g.umount_all()
-  except Exception as e:
-    logging.debug(str(e))
-    logging.warn('Unmount failed. Continuing anyway.')
 
 
 def CommonRoutines(g):
